Public forms, form editor and tags
... all work!

It's now possible to add unlimited custom fields with a lot of options for each field type. These fields become available when editing a recipient in the backend.

... and from the frontend!

Because now there's also short-code support for adding a recipient, updating the recipient, unsubscribing and also sending SMS'es straight from the frontend (all created due to popular demand).