List of group IDs now with quotes
Previously the way IDs were written in the CF7 shortcode for the groups selector, broke in CF7 when there were only IDs and no other properties. Now the IDs can be in quotes as well, which is the proper CF7-way, though the new approach is fully backwards-compable. The generator though will use the new approach, so the backwards compability is purely for forms which has not had the tags upgraded.