Pairing works + general security improvement
The pairing on initial log in of an account with no phone linking, now works and is pretty hardened in terms of brute force attempts.

Also, one line has been added to the top of every php-file just as a good measure, in case somebody tried to access any php-file directly.